+++ Fabric/src/main/java/net/potionstudios/biomeswevegone/fabric/BiomesWeveGoneFabric.java
@@ -2,10 +2,12 @@ package net.potionstudios.biomeswevegone.fabric;
 
 import corgitaco.corgilib.fabric.CorgiLibFabric;
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricDefaultAttributeRegistry;
 import net.minecraft.world.entity.SpawnPlacements;
 import net.potionstudios.biomeswevegone.BiomesWeveGone;
+import net.potionstudios.biomeswevegone.commands.BWGReloadCommand;
 import net.potionstudios.biomeswevegone.world.entity.BWGEntities;
 
 /**
@@ -38,5 +40,6 @@ public class BiomesWeveGoneFabric implements ModInitializer {
         BiomesWeveGone.commonSetup();
         BiomesWeveGone.postInit();
         ServerLifecycleEvents.SERVER_STARTING.register(BiomesWeveGone::serverStart);
+        CommandRegistrationCallback.EVENT.register((dispatcher, registryAccess, environment) -> BWGReloadCommand.register(dispatcher::register));
     }
 }