+++ Forge/src/main/java/net/potionstudios/biomeswevegone/forge/datagen/generators/LangGenerator.java
@@ -89,6 +89,8 @@ public class LangGenerator extends LanguageProvider {
         add(death("inQuicksand"), "%s tried to swim in the desert");
         add(death("cattailExplosion"), "%s got too curious and put a cattail in a campfire");
 
+        add("biomeswevegone.commands.reload.spawn.success", "Successfully reloaded Mob Spawn config");
+
         add("tooltip.waila.hatch_chance", "Hatch Chance: %s");
     }